Reorganize config remap migrations
