+++ photon-core/src/main/java/org/photonvision/vision/camera/PVCameraInfo.java
@@ -18,7 +18,12 @@
 package org.photonvision.vision.camera;
 
 import io.avaje.jsonb.Json;
+import io.avaje.jsonb.JsonType;
+import io.avaje.jsonb.Jsonb;
+import io.avaje.jsonb.Types;
 import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Objects;
 import org.wpilib.vision.camera.UsbCameraInfo;
 
@@ -55,7 +60,7 @@ public sealed interface PVCameraInfo {
      *
      * @return The unique path of the camera
      */
-    @Json.Property(value = "uniquePath")
+    @Json.Property("uniquePath")
     String uniquePath();
 
     String[] otherPaths();
@@ -281,5 +286,25 @@ public sealed interface PVCameraInfo {
 
     public static PVCameraInfo fromFileInfo(String path, String baseName) {
         return new PVFileCameraInfo(path, baseName);
+    }
+
+    public static String remapConfigJson(String configJson, String cameraType) {
+        final JsonType<Map<String, Object>> objMapJsonb =
+                Jsonb.instance().type(Types.mapOf(Object.class));
+
+        Map<String, Object> cameraMigration = objMapJsonb.fromJson(configJson);
+
+        @SuppressWarnings("unchecked")
+        var cameraMigrationIn = (Map<String, Object>) cameraMigration.get("matchedCameraInfo");
+
+        @SuppressWarnings("unchecked")
+        var cameraData = (Map<String, Object>) cameraMigrationIn.get(cameraType);
+
+        Map<String, Object> cameraMigrationOut = new HashMap<>();
+        cameraMigrationOut.putAll(cameraData);
+        cameraMigrationOut.put("type", "PVCameraInfo." + cameraType);
+
+        cameraMigration.put("matchedCameraInfo", cameraMigrationOut);
+        return objMapJsonb.toJson(cameraMigration);
     }
 }
